A library to do shell-like commands easily from Go.
For man simple shell-like tasks Go's subprocess module
is overkill, so this makes excuting shell commands easy.

I'm ambivalent about this day, because the program doesn't do
anything more than the previous day, and it feels a little
bit like mental masturbation, and a bit of exactly what I'm
trying to avoid in this project.